Move testcontainers tests under common package
