+++ dataframe-jdbc/src/test/kotlin/org/jetbrains/kotlinx/dataframe/io/testcontainers/mysqlTest.kt
@@ -1,6 +1,7 @@
-package org.jetbrains.kotlinx.dataframe.io.local
+package org.jetbrains.kotlinx.dataframe.io.testcontainers
 
 import io.kotest.matchers.shouldBe
+import kotlinx.datetime.LocalDateTime
 import org.intellij.lang.annotations.Language
 import org.jetbrains.kotlinx.dataframe.DataFrame
 import org.jetbrains.kotlinx.dataframe.annotations.DataSchema
@@ -17,10 +18,9 @@ import org.junit.AfterClass
 import org.junit.BeforeClass
 import org.junit.Ignore
 import org.junit.Test
-import org.testcontainers.mariadb.MariaDBContainer
+import org.testcontainers.mysql.MySQLContainer
 import java.math.BigDecimal
 import java.math.BigInteger
-import java.sql.Blob
 import java.sql.Connection
 import java.sql.DriverManager
 import java.sql.SQLException
@@ -28,18 +28,19 @@ import java.util.Date
 import kotlin.reflect.typeOf
 import kotlin.time.Instant
 
-// Available image tags: https://hub.docker.com/_/mariadb/tags
-private const val MARIADB_IMAGE = "mariadb:11"
+// Available image tags: https://hub.docker.com/_/mysql/tags
+private const val MYSQL_IMAGE = "mysql:8.0"
 private const val USER_NAME = "root"
 private const val PASSWORD = "pass"
 private const val TEST_DATABASE_NAME = "testKDFdatabase"
+private const val TIMESTAMP = 1726246245460
 
 @DataSchema
-interface Table1MariaDb {
+interface Table1MySql {
     val id: Int
     val bitCol: Boolean
     val tinyintCol: Int
-    val smallintCol: Short?
+    val smallintCol: Int
     val mediumintCol: Int
     val mediumintUnsignedCol: Int
     val integerCol: Int
@@ -68,11 +69,10 @@ interface Table1MariaDb {
     val enumCol: String
     val setCol: Char
     val bigintUnsignedCol: BigInteger
-    val jsonCol: String
 }
 
 @DataSchema
-interface Table2MariaDb {
+interface Table2MySql {
     val id: Int
     val bitCol: Boolean?
     val tinyintCol: Int?
@@ -109,21 +109,16 @@ interface Table2MariaDb {
 }
 
 @DataSchema
-interface Table3MariaDb {
+interface Table3MySql {
     val id: Int
     val enumCol: String
     val setCol: Char?
 }
 
-private const val JSON_STRING =
-    "{\"details\": {\"foodType\": \"Pizza\", \"menu\": \"https://www.loumalnatis.com/our-menu\"}, \n" +
-        "     \t\"favorites\": [{\"description\": \"Pepperoni deep dish\", \"price\": 18.75}, \n" +
-        "{\"description\": \"The Lou\", \"price\": 24.75}]}"
-
 @Ignore
-class MariadbTest {
+class MySqlTest {
     companion object {
-        private val mariadb: MariaDBContainer = MariaDBContainer(MARIADB_IMAGE).apply {
+        private val mysql: MySQLContainer = MySQLContainer(MYSQL_IMAGE).apply {
             withUsername(USER_NAME)
             withPassword(PASSWORD)
         }
@@ -131,12 +126,12 @@ class MariadbTest {
         private lateinit var connection: Connection
 
         private val rootUrl: String
-            get() = "jdbc:mariadb://${mariadb.host}:${mariadb.firstMappedPort}"
+            get() = "jdbc:mysql://${mysql.host}:${mysql.firstMappedPort}"
 
         @BeforeClass
         @JvmStatic
         fun setUpClass() {
-            mariadb.start()
+            mysql.start()
             connection = DriverManager.getConnection(rootUrl, USER_NAME, PASSWORD)
 
             connection.createStatement().use { st ->
@@ -162,7 +157,7 @@ class MariadbTest {
                 id INT AUTO_INCREMENT PRIMARY KEY,
                 bitCol BIT NOT NULL,
                 tinyintCol TINYINT NOT NULL,
-                smallintCol SMALLINT,
+                smallintCol SMALLINT NOT NULL,
                 mediumintCol MEDIUMINT NOT NULL,
                 mediumintUnsignedCol MEDIUMINT UNSIGNED NOT NULL,
                 integerCol INTEGER NOT NULL,
@@ -191,10 +186,12 @@ class MariadbTest {
                 enumCol ENUM('Value1', 'Value2', 'Value3') NOT NULL,
                 setCol SET('Option1', 'Option2', 'Option3') NOT NULL,
                 bigintUnsignedCol BIGINT UNSIGNED NOT NULL,
-                jsonCol JSON NOT NULL
-                CHECK (JSON_VALID(jsonCol))
+                location GEOMETRY,
+                data JSON
+                CHECK (JSON_VALID(data))
             )
-        """
+            """
+
             connection.createStatement().execute(createTableQuery.trimIndent())
 
             @Language("SQL")
@@ -231,9 +228,13 @@ class MariadbTest {
                 longtextCol LONGTEXT,
                 enumCol ENUM('Value1', 'Value2', 'Value3'),
                 setCol SET('Option1', 'Option2', 'Option3'),
-                bigintUnsignedCol BIGINT UNSIGNED
+                bigintUnsignedCol BIGINT UNSIGNED,
+                location GEOMETRY,
+                data JSON
+                CHECK (JSON_VALID(data))
             )
             """
+
             connection.createStatement().execute(createTableQuery2.trimIndent())
 
             @Language("SQL")
@@ -243,8 +244,8 @@ class MariadbTest {
                     bitCol, tinyintCol, smallintCol, mediumintCol, mediumintUnsignedCol, integerCol, intCol, 
                     integerUnsignedCol, bigintCol, floatCol, doubleCol, decimalCol, dateCol, datetimeCol, timestampCol,
                     timeCol, yearCol, varcharCol, charCol, binaryCol, varbinaryCol, tinyblobCol, blobCol,
-                    mediumblobCol, longblobCol, textCol, mediumtextCol, longtextCol, enumCol, setCol, bigintUnsignedCol, jsonCol
-                ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
+                    mediumblobCol, longblobCol, textCol, mediumtextCol, longtextCol, enumCol, setCol, bigintUnsignedCol, location, data
+                ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ST_GeomFromText('POINT(1 1)'), ?)
                 """.trimIndent()
 
             @Language("SQL")
@@ -254,8 +255,8 @@ class MariadbTest {
                     bitCol, tinyintCol, smallintCol, mediumintCol, mediumintUnsignedCol, integerCol, intCol, 
                     integerUnsignedCol, bigintCol, floatCol, doubleCol, decimalCol, dateCol, datetimeCol, timestampCol,
                     timeCol, yearCol, varcharCol, charCol, binaryCol, varbinaryCol, tinyblobCol, blobCol,
-                    mediumblobCol, longblobCol, textCol, mediumtextCol, longtextCol, enumCol, setCol, bigintUnsignedCol
-                ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
+                    mediumblobCol, longblobCol, textCol, mediumtextCol, longtextCol, enumCol, setCol, bigintUnsignedCol, location, data
+                ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ST_GeomFromText('POINT(1 1)'), ?)
                 """.trimIndent()
 
             connection.prepareStatement(insertData1).use { st ->
@@ -273,10 +274,10 @@ class MariadbTest {
                     st.setFloat(10, i * 10.0f)
                     st.setDouble(11, i * 10.0)
                     st.setBigDecimal(12, BigDecimal(i * 10))
-                    st.setDate(13, java.sql.Date(System.currentTimeMillis()))
-                    st.setTimestamp(14, java.sql.Timestamp(System.currentTimeMillis()))
-                    st.setTimestamp(15, java.sql.Timestamp(System.currentTimeMillis()))
-                    st.setTime(16, java.sql.Time(System.currentTimeMillis()))
+                    st.setDate(13, java.sql.Date(TIMESTAMP))
+                    st.setTimestamp(14, java.sql.Timestamp(TIMESTAMP))
+                    st.setTimestamp(15, java.sql.Timestamp(TIMESTAMP))
+                    st.setTime(16, java.sql.Time(TIMESTAMP))
                     st.setInt(17, 2023)
                     st.setString(18, "varcharValue$i")
                     st.setString(19, "charValue$i")
@@ -292,8 +293,7 @@ class MariadbTest {
                     st.setString(29, "Value$i")
                     st.setString(30, "Option$i")
                     st.setObject(31, BigInteger.valueOf((i * 1000).toLong()))
-                    st.setString(32, JSON_STRING)
-
+                    st.setString(32, "{\"key\": \"value\"}")
                     st.executeUpdate()
                 }
             }
@@ -313,10 +313,10 @@ class MariadbTest {
                     st.setFloat(10, i * 20.0f)
                     st.setDouble(11, i * 20.0)
                     st.setBigDecimal(12, BigDecimal(i * 20))
-                    st.setDate(13, java.sql.Date(System.currentTimeMillis()))
-                    st.setTimestamp(14, java.sql.Timestamp(System.currentTimeMillis()))
-                    st.setTimestamp(15, java.sql.Timestamp(System.currentTimeMillis()))
-                    st.setTime(16, java.sql.Time(System.currentTimeMillis()))
+                    st.setDate(13, java.sql.Date(TIMESTAMP))
+                    st.setTimestamp(14, java.sql.Timestamp(TIMESTAMP))
+                    st.setTimestamp(15, java.sql.Timestamp(TIMESTAMP))
+                    st.setTime(16, java.sql.Time(TIMESTAMP))
                     st.setInt(17, 2023)
                     st.setString(18, "varcharValue$i")
                     st.setString(19, "charValue$i")
@@ -332,6 +332,7 @@ class MariadbTest {
                     st.setString(29, "Value$i")
                     st.setString(30, "Option$i")
                     st.setObject(31, BigInteger.valueOf((i * 2000).toLong()))
+                    st.setString(32, "{\"key\": \"value\"}")
                     st.executeUpdate()
                 }
             }
@@ -348,32 +349,32 @@ class MariadbTest {
             } catch (e: SQLException) {
                 e.printStackTrace()
             }
-            mariadb.stop()
+            mysql.stop()
         }
     }
 
     @Test
     fun `basic test for reading sql tables`() {
-        val df1 = DataFrame.readSqlTable(connection, "table1").cast<Table1MariaDb>()
+        val df1 = DataFrame.readSqlTable(connection, "table1").cast<Table1MySql>()
         val result = df1.filter { "id"<Int>() == 1 }
         result[0][26] shouldBe "textValue1"
-        val byteArray = "tinyblobValue".toByteArray()
-        result[0][22] shouldBe byteArray
+        result[0][22] shouldBe "tinyblobValue".toByteArray()
 
         val schema = DataFrameSchema.readSqlTable(connection, "table1")
         schema.columns["id"]!!.type shouldBe typeOf<Int>()
         schema.columns["textCol"]!!.type shouldBe typeOf<String>()
-        schema.columns["varbinaryCol"]!!.type shouldBe typeOf<ByteArray>()
-        schema.columns["binaryCol"]!!.type shouldBe typeOf<ByteArray>()
-        schema.columns["longblobCol"]!!.type shouldBe typeOf<Blob>()
-        schema.columns["tinyblobCol"]!!.type shouldBe typeOf<Blob>()
         schema.columns["dateCol"]!!.type shouldBe typeOf<Date>()
-        schema.columns["datetimeCol"]!!.type shouldBe typeOf<Instant>()
+        schema.columns["datetimeCol"]!!.type shouldBe typeOf<LocalDateTime>()
         schema.columns["timestampCol"]!!.type shouldBe typeOf<Instant>()
         schema.columns["timeCol"]!!.type shouldBe typeOf<java.sql.Time>()
         schema.columns["yearCol"]!!.type shouldBe typeOf<Date>()
+        schema.columns["textCol"]!!.type shouldBe typeOf<String>()
+        schema.columns["varbinaryCol"]!!.type shouldBe typeOf<ByteArray>()
+        schema.columns["binaryCol"]!!.type shouldBe typeOf<ByteArray>()
+        schema.columns["longblobCol"]!!.type shouldBe typeOf<ByteArray>()
+        schema.columns["tinyblobCol"]!!.type shouldBe typeOf<ByteArray>()
 
-        val df2 = DataFrame.readSqlTable(connection, "table2").cast<Table2MariaDb>()
+        val df2 = DataFrame.readSqlTable(connection, "table2").cast<Table2MySql>()
         val result2 = df2.filter { "id"<Int>() == 1 }
         result2[0][26] shouldBe null
 
@@ -395,7 +396,7 @@ class MariadbTest {
             JOIN table2 t2 ON t1.id = t2.id
             """.trimIndent()
 
-        val df = DataFrame.readSqlQuery(connection, sqlQuery = sqlQuery).cast<Table3MariaDb>()
+        val df = DataFrame.readSqlQuery(connection, sqlQuery = sqlQuery).cast<Table3MySql>()
         val result = df.filter { "id"<Int>() == 1 }
         result[0][2] shouldBe "Option1"
 
@@ -407,18 +408,16 @@ class MariadbTest {
 
     @Test
     fun `read from all tables`() {
-        val dataframes = DataFrame.readAllSqlTables(connection, TEST_DATABASE_NAME, 1000).values.toList()
+        val dataframes = DataFrame.readAllSqlTables(connection).values.toList()
 
-        val table1Df = dataframes[0].cast<Table1MariaDb>()
+        val table1Df = dataframes[0].cast<Table1MySql>()
 
         table1Df.rowsCount() shouldBe 3
         table1Df.filter { "integerCol"<Int>() > 100 }.rowsCount() shouldBe 2
         table1Df[0][11] shouldBe 10.0
         table1Df[0][26] shouldBe "textValue1"
-        table1Df[0][31] shouldBe BigInteger.valueOf(1000L)
-        table1Df[0][32] shouldBe JSON_STRING // TODO: https://github.com/Kotlin/dataframe/issues/462
 
-        val table2Df = dataframes[1].cast<Table2MariaDb>()
+        val table2Df = dataframes[1].cast<Table2MySql>()
 
         table2Df.rowsCount() shouldBe 3
         table2Df.filter {
@@ -430,17 +429,16 @@ class MariadbTest {
 
     @Test
     fun `reading numeric types`() {
-        val df1 = DataFrame.readSqlTable(connection, "table1").cast<Table1MariaDb>()
+        val df1 = DataFrame.readSqlTable(connection, "table1").cast<Table1MySql>()
 
-        val result = df1.select("tinyintCol")
-            .add("tinyintCol2") { "tinyintCol"<Int>() }
+        val result = df1.select("tinyintCol").add("tinyintCol2") { "tinyintCol"<Int>() }
 
-        result[0][1] shouldBe 1
+        result[0][1] shouldBe 1.toByte()
 
         val result1 = df1.select("smallintCol")
-            .add("smallintCol2") { "smallintCol"<Short?>() }
+            .add("smallintCol2") { "smallintCol"<Int?>() }
 
-        result1[0][1] shouldBe 10
+        result1[0][1] shouldBe 10.toShort()
 
         val result2 = df1.select("mediumintCol")
             .add("mediumintCol2") { "mediumintCol"<Int>() }
@@ -485,7 +483,7 @@ class MariadbTest {
         val schema = DataFrameSchema.readSqlTable(connection, "table1")
 
         schema.columns["tinyintCol"]!!.type shouldBe typeOf<Int>()
-        schema.columns["smallintCol"]!!.type shouldBe typeOf<Short?>()
+        schema.columns["smallintCol"]!!.type shouldBe typeOf<Int>()
         schema.columns["mediumintCol"]!!.type shouldBe typeOf<Int>()
         schema.columns["mediumintUnsignedCol"]!!.type shouldBe typeOf<Int>()
         schema.columns["integerUnsignedCol"]!!.type shouldBe typeOf<Long>()
@@ -494,6 +492,11 @@ class MariadbTest {
         schema.columns["floatCol"]!!.type shouldBe typeOf<Float>()
         schema.columns["doubleCol"]!!.type shouldBe typeOf<Double>()
         schema.columns["decimalCol"]!!.type shouldBe typeOf<BigDecimal>()
+        // TODO: all unsigned types
+        // TODO: new mapping system based on class names
+        // validation after mapping in getObject
+        // getObject(i+1, type) catch getObject catch getString
+        // add direct mapping to getString and other methods
     }
 
     @Test
