+++ dataframe-jdbc/src/test/kotlin/org/jetbrains/kotlinx/dataframe/io/testcontainers/postgresConnectionUrlTest.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.kotlinx.dataframe.io.local
+package org.jetbrains.kotlinx.dataframe.io.testcontainers
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
